Split up Chapter Router into multiple routers. Small adjustments to various endpoints.
