+++ api/chapterRouter.js
@@ -7,14 +7,13 @@ const chaptersVolunteersDB = require("../models/chapters-volunteers-model");
 const aws_link = "https://miraclemessagesimages.s3.amazonaws.com/";
 const axios = require("axios");
 
-//TODO to be implemented
+
 const authenticationRequired = require("../middleware/Okta");
 const userInfo = require("../middleware/userInfo")
 const adminCheck = require("../middleware/Admin")
 
 
-// Returns: all chapters
-// ✔
+// Return: all chapters
 router.get("/", async (req, res) => {
   try {  
     let chapters = await chapterDB.findChapters(req.query);
@@ -40,249 +39,19 @@ router.get("/", async (req, res) => {
   }
 });
 
-//see pending chapters
-router.get("/pending", authenticationRequired, userInfo, adminCheck, async (req, res) => {
-  try { 
-    let chapters = await chapterDB.findPendingChapters();
-    console.log(chapters)
-    if(chapters.length >= 1) {
-      res.status(200).json(chapters);
-    } else {
-      res.status(201).json({"Message":"There are no pending chapters at this time"})
-    }
-  } catch {
-    res.status(500).json({
-      error: "there was a problem getting chapter",
-    });
-  }
-});
-
-
-//approve pending chapter
-router.put("/:id/approve", authenticationRequired, userInfo, async (req,res)=> {
-  const chapterId = req.params.id
-  const groups = req.jwt.claims.groups
- 
-  if(groups.includes("Admins")){
-    const chapter = await chapterDB.findBy(chapterId)
-    if(chapter.approved === false) {
-      chapterDB.approveChapter(chapterId)
-    .then(chapter => {
-      res.status(201).json({"Message": "Chapter Successfully approved"})
-    })
-    .catch(err => {
-      console.log(err)
-      res.status(500).json({err, "Message": "Something went wrong and chapter could not be approved right now"})
-    })
-    } else {
-      res.status(201).json({"Message": "Chapter was already Approved!"})
-    }
-    
-  }
-  else {
-    res.status(401).json({"Error": "User must be an admin to approve chapters"})
-  }
-
-} )
-
-//Returns: JSON a specific chapter by id
-// ✔
-router.get("/:id", async (req, res) => {
-  const chapterId = req.params.id
+//Return: a specific chapter by id
+router.get("/:chapterid", async (req, res) => {
+  const chapterId = req.params.chapterid
   try {
     const chapter = await chapterDB.findBy(chapterId);
+    chapter.volunteers = await chaptersVolunteersDB.findChapterVolunteers(chapter.id)
+    chapter.leaders = await chaptersVolunteersDB.findLeaders(chapter.id)
+
     res.status(200).json(chapter);
   } catch (error) { res.status(500).json({ message: "Error getting the chapter",});}
 });
 
-// Returns: Json of all volunteers related to a specific chapter
-// ✔
-router.get("/:id/volunteers", async (req, res) => {
-  const chapterId = req.params.id;
-  try {
-    const volunteers = await chaptersVolunteersDB.findChapterVolunteers(chapterId);
-    //Checks if there are volunteers in chapter
-    // console.log(volunteers)
-    if (volunteers.length < 1) { res.status(404).json({ message: "There is no volunteers in this chapter" });}
-    else { res.status(200).json(volunteers); }
-  } catch { res.status(500).json({ errorMessage: "There is a problem finding volunteers data" });}
-});
-
-
-router.get("/:id/pendingVolunteers", authenticationRequired, userInfo,  adminCheck, async (req, res) => {
-  const chapterId = req.params.id;
-  try {
-    const volunteers = await chaptersVolunteersDB.findPendingChapterVolunteers(chapterId);
-    //Checks if there are volunteers in chapter
-    if (volunteers.length < 1) { res.status(404).json({ message: "There are no pending volunteers for this chapter" });}
-    else { res.status(200).json(volunteers); }
-  } catch { res.status(500).json({ errorMessage: "There is a problem finding volunteers data" });}
-});
-
-
-
-router.put("/:id/approveVolunteer", authenticationRequired, userInfo,  adminCheck, async (req,res)=> {
-  const chapterId = req.params.id
-  const oktaId = req.body.oktaId
-  
-    const volunteer = await chaptersVolunteersDB.getSpecificChapterVolunteer(oktaId, chapterId)
-    if(volunteer.approved === false) {
-      chaptersVolunteersDB.approveVolunteer(oktaId, chapterId)
-    .then(volunteer => {
-      res.status(201).json({"Message": "Volunteer Successfully approved"})
-    })
-    .catch(err => {
-      console.log(err)
-      res.status(500).json({err, "Message": "Something went wrong and volunteer could not be approved right now"})
-    })
-    } else {
-      res.status(201).json({"Message": "Volunteer was already Approved!"})
-    }
-
-} )
-
-
-router.put("/:id/requestLeader", authenticationRequired, userInfo, async (req,res) => {
-  const chapterId = req.params.id
-  const oktaId = req.userInfo.sub
-  
-    const volunteer = await chaptersVolunteersDB.getSpecificChapterVolunteer(oktaId, chapterId)
-    if(volunteer.isAdmin === false && volunteer.approved === true) {
-      chaptersVolunteersDB.requestLeader(oktaId, chapterId)
-    .then(volunteer => {
-      res.status(201).json({"Message": "Volunteer Successfully Requested to become leader"})
-    })
-    .catch(err => {
-      console.log(err)
-      res.status(500).json({err, "Message": "Something went wrong and volunteer could not make the request right now"})
-    })
-    } else {
-      res.status(201).json({"Message": "Volunteer is either already an admin, or is not approved for this chapter"})
-    }
-})
-
-
-router.get("/:id/pendingLeader", authenticationRequired, userInfo, adminCheck, async (req, res) => {
-  const chapterId = req.params.id;
-  try {
-    const leaders = await chaptersVolunteersDB.findPendingChapterLeaders(chapterId);
-    if (leaders.length < 1) { res.status(404).json({ message: "There are no pending chapter leaders for this chapter" });}
-    else { res.status(200).json(leaders); }
-  } catch { res.status(500).json({ errorMessage: "There is a problem finding volunteer data" });}
-})
-
-router.put("/:id/approveLeader",  authenticationRequired, userInfo, adminCheck, async (req, res) => {
-  const chapterId = req.params.id
-  const oktaId = req.body.oktaId
-  
-    const volunteer = await chaptersVolunteersDB.getSpecificChapterVolunteer(oktaId, chapterId)
-    if(volunteer.isAdmin === false && volunteer.requestedAdmin === true) {
-      chaptersVolunteersDB.approveLeader(oktaId, chapterId)
-    .then(volunteer => {
-      res.status(201).json({"Message": "Volunteer is now a chapter leader"})
-    })
-    .catch(err => {
-      console.log(err)
-      res.status(500).json({err, "Message": "Something went wrong and volunteer could not be approved right now"})
-    })
-    } else {
-      res.status(201).json({"Message": "Volunteer is already a leader!"})
-    }
-
-})
-
-router.put("/:id/declineLeader",  authenticationRequired, userInfo, adminCheck, async (req, res) => {
-  const chapterId = req.params.id
-  const oktaId = req.body.oktaId
-  
-    const volunteer = await chaptersVolunteersDB.getSpecificChapterVolunteer(oktaId, chapterId)
-    if(volunteer.isAdmin === false && volunteer.requestedAdmin === true) {
-      chaptersVolunteersDB.declineLeader(oktaId, chapterId)
-    .then(volunteer => {
-      res.status(201).json({"Message": "Volunteer was declined ;-;"})
-    })
-    .catch(err => {
-      console.log(err)
-      res.status(500).json({err, "Message": "Something went wrong and volunteer could not be declined right now"})
-    })
-    } else {
-      res.status(201).json({"Message": "Volunteer is either currently a chapter leader, or didn't request to become a leader"})
-    }
-})
-
-
-// Returns: JSON of all reunions specific to a chapter 
-// ✔
-router.get("/:id/reunions", async (req, res) => {
-  const chapterId = req.params.id;
-  try {
-    const reunions = await reunionDB.findByChapterId(chapterId);
-    res.status(200).json(reunions);
-  } catch (error) {
-    res.status(500).json({ "Error": "Error getting the reunions", error });
-  }
-});
-
-
-
-router.get("/:id/reunion", async (req,res)=> {
-  const { reunionId } = req.body
-  try {
-    const reunion = await reunionDB.findById(reunionId)
-    res.status(200).json(reunion);
-  } catch (error) {
-    res.status(500).json({ "Error": "Error retrieving pending reunions", error})
-  }
-})
-
-router.get("/:id/pendingReunions",  authenticationRequired, userInfo, adminCheck, async (req,res)=> {
-  const chapterId = req.params.id
-  try {
-    const pending = await reunionDB.findPendingReunions(chapterId)
-    res.status(200).json(pending);
-  } catch (error) {
-    res.status(500).json({ "Error": "Error retrieving pending reunions", error})
-  }
-})
-
-router.put("/:id/approveReunion",  authenticationRequired, userInfo, adminCheck, async (req,res) => {
-  const { reunionId } = req.body
-  try {
-    const reunion = await reunionDB.findById(reunionId)
-    if (reunion.approved === false) {
-      const approved = await reunionDB.approveReunion(reunionId)
-      res.status(200).json({ "Message": "Reunion successfully approved",approved})
-    } else {
-      res.status(401).json({ "Error": "This reunion is either already approved or doesn't exist"})
-    }
-    
-  } catch (error) {
-    res.status(500).json({ "Error":"There was an error approving this reunion", error})
-  }
-
-
-})
-
-// Returns a specific volunteer to a chapter
-// ✔
-router.get("/:id/volunteer", async (req, res) => {
-  let chapterId = req.params.id;
-  let oktaId = req.body.oktaid;
-  try {
-    const isVolunteerInChapter = await chaptersVolunteersDB.getSpecificChapterVolunteer(
-      oktaId,
-      chapterId
-    );
-    res.status(200).json(isVolunteerInChapter);
-  } catch (error) {
-    res.status(500).json({
-      message: "Error getting the chapter",
-    });
-  }
-});
-
-// create new chapter
-// ✔
+//Post: create new chapter
 router.post("/", authenticationRequired, userInfo, async (req, res) => {
   const newChapter = req.body;
   newChapter.requestedBy = req.userInfo.sub
@@ -337,57 +106,9 @@ router.post("/", authenticationRequired, userInfo, async (req, res) => {
   }
 });
 
-// Creates new reunion connected to chapter
-// ✔
-router.post("/:id/reunions", authenticationRequired, userInfo, async (req, res) => {
-  
-  const newReunion = req.body
-  newReunion.chapterid = req.params.id
-  newReunion.volunteersid = req.userInfo.sub
-
-
-const reunionCoordinates = await axios
-  .get(
-    `https://api.mapbox.com/geocoding/v5/mapbox.places/${newReunion.city},${newReunion.state}.json?access_token=${process.env.MAPBOX_API}`
-  )
-  .then((res) => {
-    return res.data.features[0].geometry.coordinates;
-  })
-  .catch((err) => {
-    console.log("could not get lat & lng from mapbox", err);
-  });
-
-newReunion.latitude = reunionCoordinates[1];
-newReunion.longitude = reunionCoordinates[0];
-
-if (req.files && req.files.reunion_img) {
-  const { reunion_img }  = await req.files;
-
-  try {
-    uploadToS3(reunion_img, res);
-  } catch (error) {
-    res.status(500).json({ error: "error uploading the image to AWS" });
-  }
-
-  // storing the reunion image url in the newChapter object:
-  const reunionImgName = await req.files.reunion_img.name;
-  const encodedReunionImgName = encodeURI(reunionImgName);
-  newReunion.reunion_img = aws_link + encodedReunionImgName;
-}
-  try {
-    const id = await reunionDB.addReunion(newReunion);
-
-    res.status(200).json(id);
-  } catch (error) {
-    res
-      .status(500)
-      .json({ errorMessage: "error submitting reunion to the chapter", error });
-  }
-});
-
-//registers a user to a chapter to await approval
-router.post("/:id/volunteer", authenticationRequired, userInfo, async (req, res) => {
-  let chapterId = req.params.id;
+//Post: register a user to a chapter to await approval
+router.post("/:chapterId/register", authenticationRequired, userInfo, async (req, res) => {
+  let chapterId = req.params.chapterId;
   let oktaId = req.userInfo.sub;
   try {
     const isVolunteerInChapter = await chaptersVolunteersDB.getSpecificChapterVolunteer(
@@ -402,7 +123,7 @@ router.post("/:id/volunteer", authenticationRequired, userInfo, async (req, res)
       );
 
       res.status(201).json({
-        message: `You have successfully signed up for this chapter.`,
+        "message": `You have successfully signed up for this chapter.`,
         id: signedUp,
       });
     } else {
@@ -419,12 +140,32 @@ router.post("/:id/volunteer", authenticationRequired, userInfo, async (req, res)
   }
 });
 
-//update chapter info
-// ✔
-router.put("/:id",authenticationRequired, userInfo, adminCheck, async (req, res) => {
+
+//Put: Request to become leader of chapter
+router.put("/:id/requestLeader", authenticationRequired, userInfo, async (req,res) => {
+  const chapterId = req.params.id
+  const oktaId = req.userInfo.sub
+  
+    const volunteer = await chaptersVolunteersDB.getSpecificChapterVolunteer(oktaId, chapterId)
+    if(volunteer.isAdmin === false && volunteer.approved === true) {
+      chaptersVolunteersDB.requestLeader(oktaId, chapterId)
+    .then(volunteer => {
+      res.status(201).json({"Message": "Volunteer Successfully Requested to become leader"})
+    })
+    .catch(err => {
+      console.log(err)
+      res.status(500).json({err, "Message": "Something went wrong and volunteer could not make the request right now"})
+    })
+    } else {
+      res.status(201).json({"Message": "Volunteer is either already an admin, or is not approved for this chapter"})
+    }
+})
+
+//Update: chapter info
+router.put("/:chapterId",authenticationRequired, userInfo, adminCheck, async (req, res) => {
   try {
     const updatedChapter = req.body;
-    const current = await chapterDB.findBy(req.params.id)
+    const current = await chapterDB.findBy(req.params.chapterId)
     if(updatedChapter.city && updatedChapter.state) {
       const chapterCoordinates = await axios
     .get(
@@ -463,7 +204,7 @@ router.put("/:id",authenticationRequired, userInfo, adminCheck, async (req, res)
       || updatedChapter.email){
       
       const chapter = await chapterDB.updateChapter(
-          req.params.id,
+          req.params.chapterId,
           updatedChapter,
           current
         )
@@ -482,108 +223,13 @@ router.put("/:id",authenticationRequired, userInfo, adminCheck, async (req, res)
 
 });
 
-//update reunion
-router.put("/:id/reunion",authenticationRequired, userInfo, adminCheck, async (req, res) => {
-  try {
-    const updatedReunion = req.body;
-    const current = await reunionDB.findById(req.params.id)
-    if(updatedReunion.city && updatedReunion.state) {
-      const reunionCoordinates = await axios
-    .get(
-      `https://api.mapbox.com/geocoding/v5/mapbox.places/${updatedReunion.city},${updatedReunion.state}.json?access_token=${process.env.MAPBOX_API}`
-    )
-    .then((res) => {
-      return res.data.features[0].geometry.coordinates;
-    })
-    .catch((err) => {
-      console.log("could not get lat & lng from mapbox", err)
-      res.status(500).json({"Error": "could not get lat & lng from mapbox", err})
-    });
-      
-      updatedReunion.latitude = reunionCoordinates[1];
-      updatedReunion.longitude = reunionCoordinates[0];
-    }
-
-    if (req.files && req.files.reunion_img) {
-      const  { reunion_img }  = await req.files;
-      try {
-        uploadToS3(reunion_img, res);
-      } catch (error) {
-        res.status(500).json({ error: "error uploading the image to AWS" });
-      }
-      const reunionImgName = await req.files.reunion_img.name;
-      const encodedReunionImgName = encodeURI(reunionImgName);
-      updatedReunion.reunion_img = aws_link + encodedReunionImgName;
-    }
-   
-    if(updatedReunion.longitude 
-      || updatedReunion.title 
-      || updatedReunion.reunion_img 
-      || updatedReunion.title 
-      || updatedReunion.story 
-      || updatedReunion.link_to_media){
-      
-      const reunion = await reunionDB.updateReunion(
-          req.body.reunionId,
-          updatedReunion,
-          current
-        )
-        res.status(200).json(reunion);
-      }
-    else {
-      res.status(401).json({"Error": "Please Submit something to change"})
-    }
-  } catch (error) {
-    console.log(error)
-    res.status(500).json({
-      "Message": "Error updating the reunion",
-      error
-    });
-  }
-
-});
-//deletes chapter from db
-router.delete("/:id", authenticationRequired, userInfo, adminCheck, (req, res) => {
-    const chapterId = req.params.id;
+//Delete: chapter from db
+router.delete("/:chapterId", authenticationRequired, userInfo, adminCheck, (req, res) => {
+    const chapterId = req.params.chapterId;
       chapterDB.removeChapter(chapterId)
       .then( chapter => {res.status(200).json({ "Message": "Chapter successfully deleted."})})
       .catch(err => {res.status(500).json({"Error": "There was an error deleting this chapter"})})
 
 });
-
-//Delete a volunteer from a specific chapter 
-router.delete("/:id/volunteer", authenticationRequired, adminCheck, async (req, res) => {
-  const chapterId = req.params.id;
-  const oktaId = req.body.oktaid;
-    try {
-     const count = await chaptersVolunteersDB.removeSpecificChapterVolunteer(
-        oktaId,
-       chapterId
-     );
-
-     res.status(200).json(count);
-    } catch (error) {
-     res.status(500).json({ errorMessage: "error removing volunteer" });
-    }
-});
-
-//deletes a reunion registered to specified chapter
-router.delete("/:id/reunion",authenticationRequired, adminCheck, async (req, res) => {
-  const { reunionId } = req.body
- 
-    const reunionSpecified = await reunionDB.findById(reunionId)
-    console.log(reunionSpecified.length)
-    if(reunionSpecified.length >= 1){
-      reunionDB.remove(reunionId)
-      .then(del => {res.status(201).json({"Message":"Reunion Successfully Deleted"})})
-      .catch(err => {
-        console.log(err)
-        res.status(500).json({"Error": "Something went wrong", err})
-      })
-    } else {
-      res.status(401).json({"Error":"Reunion doesn't exist to delete"})
-    }
-  
-} )
 
 module.exports = router;
