+++ middleware/userInfo.js
@@ -1,18 +1,21 @@
-const Axios = require ("axios")
+const Axios = require("axios")
 
 
 
 
 function userInfo (req,res,next) {
+        console.log("hello")
     const config = {headers: {Authorization: "Bearer " + req.accessToken}}
     Axios.get("https://dev-750287.okta.com/oauth2/default/v1/userinfo", config)
     .then(res => {
-        // console.log(res.data)
+        
         req.userInfo = res.data
+
+        
         next()
     })
     .catch(error => {
-        next(error)
+        res.status(401).json({"Error":"Cannot Validate user", error})
     })
 }
