+++ server.js
@@ -5,15 +5,13 @@ const morgan = require("morgan");
 const fileupload = require("express-fileupload");
 const cors = require("cors");
 const helmet = require("helmet");
-const authenticationRequired = require("./middleware/Okta.js");
 
-//TODO return and test to see if this route is still needed
-// const uploadRouter = require("./api/uploadRouter");
 
 const chaptersRouter = require("./api/chapterRouter.js");
 const reunionRouter = require("./api/reunionRouter");
-
+const volunteerRouter = require("./api/volunteerRouter")
 const userRouter = require("./api/userRouter.js")
+const pendingRouter = require("./api/pendingRouter")
 
 server.use(helmet());
 server.use(cors());
@@ -28,6 +26,8 @@ server.get("/", (req, res) => {
 server.use("/api/user", userRouter);
 server.use("/api/chapter", chaptersRouter);
 server.use("/api/reunion", reunionRouter);
+server.use("/api/volunteer", volunteerRouter);
+server.use("/api/pending", pendingRouter)
 
 // custom logging function 
 function logger(req, res, next) {
